+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 3.39.1
+* Minor bug fix.
+
+## 3.39.0
+* Added GetLicenseEntitlementSetTier() function which returns the tier of the license entitlement set.
+
+## 3.38.0
+* Added support for migration of activation data from LA_USER and LA_SYSTEM to LA_ALL_USERS for macOS and Windows.
+
 ## 3.37.0
 * Added expiresAt to Feature Entitlements.
 * Added totalActivations and totalDeactivations to User License.